Log real renew date on each run
Previously, the renew_at log message would be one month after the
current date instead of one month prior to the cert expiration date.
That meant it would always print out a different date each day. This
changes the logic to store the renewal date and compare that with the
current date, always printing the same renewal date.